pyproject.toml
[tool.poetry]
name = "yubikey-oidc"
version = "0.1.0"
description = "Log into AWS using Yubikey and OIDC"
authors = ["Leandro Lima <leandro@lls-software.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
yubid = "yubikey_oidc.cli:cli"

[tool.poetry.dependencies]
python = "^3.12"
PyJWT = "^2.8.0"
boto3 = "^1.34.117"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
